Create Updated Human Version JSON V2
The "symptom" intents from previous version were removed since it intervered much with the catch-all intent causing errors. What actualy worked suprisingly well, because of those intents was the amount of times you were asked to elobarate, however it almost complete ignored the filler words, which was the point of this study.